+++ disco/handler.go
@@ -10,13 +10,14 @@ import (
 	"mellium.im/xmlstream"
 	"mellium.im/xmpp/disco/info"
 	"mellium.im/xmpp/disco/items"
+	"mellium.im/xmpp/form"
 	"mellium.im/xmpp/mux"
 	"mellium.im/xmpp/stanza"
 )
 
 // Handle returns an option that configures a multiplexer to handle service
 // discovery requests by iterating over its own handlers and checking if they
-// implement info.FeatureIter or items.Iter.
+// implement info.FeatureIter, info.IdentityIter, form.Iter, or items.Iter.
 func Handle() mux.Option {
 	return func(m *mux.ServeMux) {
 		h := &discoHandler{ServeMux: m}
@@ -26,7 +27,7 @@ func Handle() mux.Option {
 }
 
 type discoHandler struct {
-	*mux.ServeMux
+	ServeMux *mux.ServeMux
 }
 
 func (h *discoHandler) HandleXMPP(t xmlstream.TokenReadEncoder, start *xml.StartElement) error {
@@ -46,13 +47,39 @@ func (h *discoHandler) HandleIQ(iq stanza.IQ, r xmlstream.TokenReadEncoder, star
 		}
 		switch start.Name.Space {
 		case NSInfo:
-			pw.CloseWithError(h.ServeMux.ForFeatures(node, func(f info.Feature) error {
+			err := h.ServeMux.ForFeatures(node, func(f info.Feature) error {
 				_, ok := seen[f.Var]
 				if ok {
 					return nil
 				}
 				seen[f.Var] = struct{}{}
 				_, err := xmlstream.Copy(pw, f.TokenReader())
+				return err
+			})
+			if err != nil {
+				pw.CloseWithError(err)
+				return
+			}
+			for k := range seen {
+				delete(seen, k)
+			}
+			err = h.ServeMux.ForIdentities(node, func(f info.Identity) error {
+				loopKey := f.Category + ":" + f.Type + ":" + f.Name + ":" + f.Lang
+				_, ok := seen[loopKey]
+				if ok {
+					return nil
+				}
+				seen[loopKey] = struct{}{}
+				_, err := xmlstream.Copy(pw, f.TokenReader())
+				return err
+			})
+			if err != nil {
+				pw.CloseWithError(err)
+				return
+			}
+			pw.CloseWithError(h.ServeMux.ForForms(node, func(f *form.Data) error {
+				result, _ := f.Submit()
+				_, err := xmlstream.Copy(pw, result)
 				return err
 			}))
 		case NSItems:
@@ -73,8 +100,4 @@ func (h *discoHandler) HandleIQ(iq stanza.IQ, r xmlstream.TokenReadEncoder, star
 		*start,
 	)))
 	return err
-}
-
-func (*discoHandler) ForItems(string, func(items.Item) error) error {
-	return nil
 }