+++ mux/mux.go
@@ -19,6 +19,7 @@ import (
 	"mellium.im/xmpp"
 	"mellium.im/xmpp/disco/info"
 	"mellium.im/xmpp/disco/items"
+	"mellium.im/xmpp/form"
 	"mellium.im/xmpp/internal/ns"
 	"mellium.im/xmpp/stanza"
 )
@@ -278,6 +279,82 @@ func (m *ServeMux) ForFeatures(node string, f func(info.Feature) error) error {
 	for _, h := range m.presencePatterns {
 		if featureIter, ok := h.(info.FeatureIter); ok {
 			err := featureIter.ForFeatures(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+// ForIdentities implements info.IdentityIter for the mux by iterating over
+// all child handlers.
+func (m *ServeMux) ForIdentities(node string, f func(info.Identity) error) error {
+	for _, h := range m.patterns {
+		if identIter, ok := h.(info.IdentityIter); ok {
+			err := identIter.ForIdentities(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.iqPatterns {
+		if identIter, ok := h.(info.IdentityIter); ok {
+			err := identIter.ForIdentities(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.msgPatterns {
+		if identIter, ok := h.(info.IdentityIter); ok {
+			err := identIter.ForIdentities(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.presencePatterns {
+		if identIter, ok := h.(info.IdentityIter); ok {
+			err := identIter.ForIdentities(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+// ForIdentities implements form.Iter for the mux by iterating over all child
+// handlers.
+func (m *ServeMux) ForForms(node string, f func(*form.Data) error) error {
+	for _, h := range m.patterns {
+		if formIter, ok := h.(form.Iter); ok {
+			err := formIter.ForForms(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.iqPatterns {
+		if formIter, ok := h.(form.Iter); ok {
+			err := formIter.ForForms(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.msgPatterns {
+		if formIter, ok := h.(form.Iter); ok {
+			err := formIter.ForForms(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.presencePatterns {
+		if formIter, ok := h.(form.Iter); ok {
+			err := formIter.ForForms(node, f)
 			if err != nil {
 				return err
 			}